requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
llama-index
langchain==1.2.0
langchain-core==1.2.6
groq==0.9.0
streamlit==1.52.2
sqlalchemy==2.0.45
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.5
